requirements.txt
Mastodon.py==1.8.1
tox==4.14.1
feedparser==6.0.11